requirements.txt
flask
flask-cors
flask_apscheduler
pillow
